More refactor for fabric api extensions
