+++ src/api/kotlin/xyz/wagyourtail/unimined/api/fabric/LegacyFabricApiExtension.kt
@@ -1,8 +1,6 @@
 package xyz.wagyourtail.unimined.api.fabric
 
 import org.gradle.api.Project
-import java.net.URL
-import javax.xml.parsers.DocumentBuilderFactory
 
 /**
  * helper class for getting legacy fabric api parts.
@@ -13,43 +11,23 @@ import javax.xml.parsers.DocumentBuilderFactory
  * ```groovy
  * dependencies {
  *    ...
- *    // fabric api part
+ *    // legacy fabric api part
  *    modImplementation legacyFabricApi.module("fabric-api-base", "0.67.1+1.19.2")
  * }
  */
 @Suppress("unused")
-open class LegacyFabricApiExtension {
+open class LegacyFabricApiExtension : FabricLikeApiExtension("legacy-fabric-api") {
     companion object {
         fun apply(target: Project) {
             target.extensions.create("legacyFabricApi", LegacyFabricApiExtension::class.java)
         }
     }
 
-    fun module(name: String, version: String): String {
-        val url = URL("https://repo.legacyfabric.net/repository/legacyfabric/net/legacyfabric/legacy-fabric-api/legacy-fabric-api/$version/legacy-fabric-api-$version.pom")
-        url.openStream().use {
-            val dbf = DocumentBuilderFactory.newInstance()
-            val db = dbf.newDocumentBuilder()
-            val doc = db.parse(it)
-            val elements = doc.getElementsByTagName("dependency")
-            for (i in 0 until elements.length) {
-                val element = elements.item(i)
-                var correct = false
-                var vers: String? = null
-                for (j in 0 until element.childNodes.length) {
-                    val child = element.childNodes.item(j)
-                    if (child.nodeName == "artifactId" && child.textContent == name) {
-                        correct = true
-                    }
-                    if (child.nodeName == "version") {
-                        vers = child.textContent
-                    }
-                }
-                if (correct) {
-                    return "net.legacyfabric.legacy-fabric-api:$name:$vers"
-                }
-            }
-        }
-        throw IllegalStateException("Could not find module $name in legacy-fabric-api $version")
+    override fun getUrl(version: String): String {
+        return "https://repo.legacyfabric.net/repository/legacyfabric/net/legacyfabric/legacy-fabric-api/legacy-fabric-api/$version/legacy-fabric-api-$version.pom"
+    }
+
+    override fun getArtifactName(moduleName: String, version: String?): String {
+        return "net.legacyfabric.legacy-fabric-api:$moduleName:$version"
     }
 }
